retrieving data from form into contract object, get correct rate from back with contract object

+++ credit_auto_Java/src/main/java/com/bnpp/creditauto/dao/RateDao.java
@@ -36,7 +36,7 @@ public class RateDao extends AbstractDao<Rate> {
 		
 		query.setParameter("cat", contract.getVehicleCategory());
 		query.setParameter("vehiclePrice", contract.getVehiclePrice());
-		query.setParameter("duration", contract.getLoanDuration());
+		query.setParameter("duration", contract.getLoanDuration().longValue());
 		return query.getSingleResult();
 	}
